requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2020.12.5
chardet==4.0.0
coards==1.0.5
Django==3.1.4
django-celery==3.3.1
django-ckeditor==6.0.0
django-crontab==0.7.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-scheduler==0.9.3
docopt==0.6.2
Extract==1.0
icalendar==4.0.7
idna==2.10
ijson==3.1.3
Jinja2==2.11.2
kombu==3.0.37
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
Pydap==3.2.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
requests==2.25.1
schedule==0.6.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
Unidecode==1.1.1
urllib3==1.26.2
urlopen==1.0.0
WebOb==1.8.6
whitenoise==5.2.0
